+++ poulpy-cpu-avx512/src/tests.rs
@@ -1,3 +1,4 @@
+#[cfg(feature = "enable-ckks")]
 mod ckks_tests;
 
 poulpy_core::core_parity_test_suite! {
@@ -103,3 +104,6 @@ poulpy_core::core_parity_test_suite! {
         glwe_tensor => poulpy_core::test_suite::parity::test_glwe_tensor_parity,
     }
 }
+
+#[cfg(all(feature = "enable-avx512f", target_arch = "x86_64", target_feature = "avx512f"))]
+poulpy_bin_fhe::bin_fhe_backend_test_suite!(mod bin_fhe_fft64, backend = crate::FFT64Avx512);